zerobounce/requirements.txt
# Core
Django==5.1.1
django-configurations==2.5.1
gunicorn==23.0.0
newrelic==9.13.0

# For the persistence stores
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Model Tools
django-model-utils==4.5.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.15.2
django-filter==24.3

# Developer Tools
ipdb==0.13.13
ipython==8.27.0
mkdocs==1.6.1
flake8==7.1.1

# Testing
mock==5.1.0
factory-boy==3.3.1
pytest-django==4.9.0
coverage==7.6.1

# Static and Media Storage
django-storages==1.14.4
boto3==1.35.10

# pyjwt
pyjwt
cryptography
django-environ

django-cors-headers
drf-spectacular

whitenoise
drf-nested-routers

pytest
pytest-django

factory-boy

celery-redbeat==2.2.0
celery[redis]==5.4.0
requests=2.32.2